Update awsEcr.md Add step to update database credentials...
The previous instruction modules do provide instruction to 
add the DB username and password to GitHub Actions Secrets.
This will be required when the deployment happens through that
channel, but it does no good for the first manual deployment.

Not performing this step will lead to a "Error occurred during operation 'ECS Deployment Circuit Breaker was triggered'."
Digging in to the issue (through multiple layers of complexities)
reveals that the instance is not able to connect to the database
which causes it to shut down and fail. After repeatedly attempting
to restart the task, the Service will eventually fail to deploy
with a less-than-helpful message.