Fix the full 40 digit SHA to use from the 3rd party action...
The 'v1' that existed already is a tag which resolves to a SHA.
This works in industry to get the right version, but a bad actor
could still reassign that tag to a different commit potentially
containing nefarious code.

By assigning the full 40 digit SHA, we prevent this kind of attack.
Tests that attempt to create a 40 digit tag (nefariously) pointing
to a different commit are foiled. The 'git' system directly uses
the commit with the full name rather than the tag.

Interestingly, this tag-redirection attack does work on abbreviated
SHAs. If a particular commit is prefixed with 0a00a74, creating a tag
with that same name does misdirect the system when performing a
'git checkout'. For this purpose, we specify the entire 40 digit SHA.