Update deliverable7Environments.md Link directly to instructions
The original links took students to the deliverable; however,
the deliverable doesn't directly contain the information needed.

Deploying a new stage does not require _all_ of the setup steps
required, but only some. The updated links directly point to
the steps that need to happen.

Other language and/or whitespace was also adjusted.