Consistently avoid ; at end of single-statement shell scripts...
Portions of the code follow this pattern, but some portions
had semi-colons at the end of the lines. Semi-colons are
required in between multiple statements on the same line.
This improves consistency across the instruction to use a
common style.

Updates were located by searching for the regex `;\s+\n`
among only *.md and *.sh files. This produced hits in
31 unique files, but many of these were JS code samples
in markdown files which were manually skipped.