Update deliverable8Metrics.md
I thought the instructions and rubric could be a bit more explicit for the need to simulate traffic. While I have graded a lot of these dashboards this semester, I have observed that quite a few students have submitted dashboards with empty visualizations, or visualizations that have all of their metrics reporting a "0" for the day before, the day of, and the day after they submitted, which undermines the whole point of the deliverable. I believe that this sets a better expectation, allowing the students to feel more confident and better prepared for the chaos testing deliverable.